Factor out globbing logic to separate module
